Add Authenticate Features using AuthRepository class
enum AuthStatus {
unknown,
authenticated,
unauthenticated 
}

class AuthState : has arguments as authStatus, user(firebase.auth)
factory -> unknown


class AuthProvider , ChangeNotifier
make a private state argument and getter

To update this widget also with the user's auth data, It should listen the data of Auth.
And also should access to the user's status(it should be able to control the user's state)
So, Make a AuthRepository argument in this class and get data from constructor when it's loaded.

89. Chapter 87. AuthState & AuthProvider ChangeNotifier